mailbox protocol: correctly handle UniPro interrupt status
Wait for the UniPro interrupt status to show a mailbox interrupt is pending
before reading the mailbox.  After clearing the mailbox to ack, wait for the
interrupt status again, and read the mailbox again, to suppress the interrupt.

Testing Done: bring up nuttx on apb2
Bug Number: Nontrivial
Review URL:
Reviewed by:

Signed-off-by: Eli Sennesh <esennesh@leaflabs.com>